pyproject.toml
[project]
name = "supervisionolt-backend-django"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.3",
    "channels>=4.3.2",
    "croniter>=6.2.2",
    "django>=6.0.3",
    "django-celery-beat>=2.9.0",
    "django-celery-results>=2.6.0",
    "django-cors-headers>=4.9.0",
    "django-debug-toolbar>=6.3.0",
    "django-environ>=0.13.0",
    "django-extensions>=4.1",
    "django-filter>=25.2",
    "django-import-export>=4.4.0",
    "django-jazzmin>=3.0.4",
    "django-redis>=6.0.0",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "drf-yasg>=1.21.15",
    "ipython>=9.12.0",
    "pandas>=3.0.2",
    "pillow>=12.2.0",
    "psycopg2-binary>=2.9.11",
    "python-dateutil>=2.9.0.post0",
    "python-decouple>=3.8",
    "redis>=7.4.0",
    "requests>=2.33.1",
    "whitenoise>=6.12.0",
]
